Move Bill events to application layer
