+++ bill/src/main/java/com/tungstun/bill/application/bill/event/BillCreated.java
@@ -1,4 +1,4 @@
-package com.tungstun.bill.port.messaging.out;
+package com.tungstun.bill.application.bill.event;
 
 public record BillCreated(
         Long id,
