+++ bill/src/main/java/com/tungstun/bill/application/bill/event/BillDeleted.java
@@ -1,4 +1,4 @@
-package com.tungstun.bill.port.messaging.out;
+package com.tungstun.bill.application.bill.event;
 
 public record BillDeleted(
         Long id,
